data export: fix error on exporting secondary rows
Queryui changes led to the main grid not always containing an id column, the
identifying column name is stored in grid.rowtarget.dataIndex instead.
This led to exports being empty as soon as it was tried to include
secondary columns, which looked up the id of their main column, but not
finding it as 'id'.