change map collapse direction
The map does not free space when collapsing to the top. Collapsing
it to the right makes more sense here. Still, the toolbar has the
collapse/expand button at the wrong end, causing it to 'jump'. As a
remedy the header itself causes collapse/expand. We'll see how users
take this.