rework printSelection printing in dynamicGrid
Visible columns (except the ones defined as 'ignored' in printSelection)
will be exported. If a custom renderer is defined in the grid, this
renderer is also used here.
Unsure if this is a good approach, as the printSelection
cannot pass a valid 'cell' argument.
Renderers using that argument will currently fail.