rudimentary splash screen
For slow connections, it is better to have some feedback at the
beginning.
The content of the index.html will be replaced as soon as extJS starts
up. So, basic html (perhaps even basic javascript) that is parsed
before loading extjs or openlayers will be shown initially. We know
nothing of the extJS framework at this time, so we cannot give any
meaningful progress or timer.
Loading the scripts not in the header, but in the body is non-standard,
but accepted by all browsers; and it helps avoiding an empty screen
before all scripts arrived.