Add copy functionality to ort window
Ort objects can now be copied at the edit ort window.
This requires the original ort to be saved und writable for the user.
The copy itself is not persisted immediately, as two ort objects may not have
completely identical attributes. See the OrtFactory Class in the Lada-Server
repository for further details, which attribute must be different between
two ort objects. This restriction is checked on the server side and by the
client, which adds error messages to the fields, that are preventing the
ort to be saved.