Add logout button to the main panel
The button starts a logout requests to the shibboleth sp.
This should trigger the logout after which the application reloads.
Note that the logout uses multiple redirects which may lead to different origins.
To prevent browsers from blocking the redirects, the idp must append the proper
headers to the response.