Add expired session handling to rest proxy
The client now recognizes an expired shibboleth sp session by receiving an
html form as a result of a lada server query.
If a form is received the client follows the redirects until the lada server
endpoint is reached. The sp session is then renewed.

Note: During this process, the initial requests JSON data is dropped.
If the initial request is a POST, PUT or DELETE request requiring payload,
it is likely to fail and should be repeated by the user.
GET requests should be successful as the URL parameters get preserved during
the process.