Reset dateTime widget changes to a working state
This reverts commit ce5dd940577d73ef238a2dab2890a90ad638c26c.
This reverts commit aeecfb9b423390bc1b91adff68eeaa857c29cfa3.

intermediate state: setting/getting of dates is correct, if the
datetime input is assumed to be in localtime. This still needs to be
changed (an input done while in utc display should be given as if it was
utc to begin with) also, the initial rendering of the hour/minute picker
seems to be nonfunctional.