Update print window to react to result grid changes
The print window is now a singleton which reacts to changes in the result grid.
This should avoid situations where one or multiple open windows still have
an old grid reference causing runtime errors.
It should now be possible to change the selection or the query itself and still
use the same print window to print reports.