Show record windows before loading data
Messung and Probe windows can now be created and shown without a record set.
The window is then created with a placeholder panel showing a loading mask. To prevent
the creation of multiple windows instances for the same record, a record id must be set.

In the query result grid, clicks on the open buttons will now use this to instantly show
the respective windows to prevent multiple clicks on the buttons in environments with high
latency or low bandwidth.