Fix tagwidget in probe create windows
If a new probe was created, the ExtJS String type id was set in the tagwidget
which caused tag requests to fail. The id should no longer be set for new probe records.

If a new probe was saved, the tags were applied before the probe record was saved. The tags should now be
saved after the new record is persisted to prevent the loss of the assigned tags.