Improve handling of filtered store
This is the same solution implemented for the Netzbetreiber widget
with rev. 15b595124. It allows filtering the selectable entries
while keeping the ability to load non-selectable entries from the
store for display of existing data (e.g. a tag associated to a
Messstelle the user cannot select when creating tags).

In passing, remove code that does not seem to add anything useful
to what's already done in the parent class.