Adapt de-/serialization of Probe objects
Set type and nullability of all fields in the model. Timestamps are
now serialized as a string representing milliseconds since epoch.

Instead of serializing empty strings as null using the same piece of
code again and again, just do not write these empty strings to the
record by passing true to getFieldValues()'s dirtyOnly argument (this
was set to false with Rev. fb00c1c65, but with no documented or
obvious reason). Also simplify the respective code and reduce
duplication.

While at it, remove some unnecessary configs that are equivalent
to defaults.