Fix initial selection in Deskriptor widgets
The Deskriptor widget prevents to be a combobox but is actually a
Lada.view.widget.base.ComboBox, which is a panel. Thus, we have to
use setStore() in order to have the store properly set also to the
combobox contained in the panel. Not doing so had the effect, that
DeskriptorFieldset.setMediaSN() was operating on empty default
stores of the comboboxes and therefore did not load or select
anything into the comboboxes. Note that due to autoLoading the
store before rev. b26024cde, the on-load handler function did add
the store to the combobox before use (accidentally?). Having it
set correctly upfront, we can save that line in the on-load handler.