Adapt de-/serialization of GridColumnValue objects
Set type and nullability of all fields in the model. Remove 'qid'
because it's unknown to the server and unused in the client.
Note that userId is 'allowNull: true' because it should be
initialized with null on creation of new instances to allow
the server to set it to the requesting user.