Do not unnecessarily create a new window
Doing so led to unexpected behavior in callbacks trying to find
records in stores that did not exist anymore, i.e. in
DeskriptorFieldset.setMediaSN(), where cbox.store.load() was
called successfully, but within the callback, cbox.store was
already null.

The pattern of creating a new window after successful creation
of a new record was most likely copied from the controller for
the Probe form. There, different windows are used for creating
respectively editing objects, which makes it necessary to
switch to the editing window after creation of the record.