Fixup rev. ef86189de5fa1c7686d11a53c02c592ec122ba93
Keeping the config at the panel leads to forms considering the
panel as a form field and calling methods on it (like validate())
that it does not have. This problem does not seem to occur with
Lada.view.widget.base.ComboBox, maybe because it extends
Ext.form.Panel, while here Ext.panel.Panel is extended
(accidentally?).