Do not store selection as a store in store
This made it impossible to have to distinct widgets (as e.g. when
opening a Messung from DynamicGrid, which opens the corresponding
Probe, too). The load spinner did not disappear for one of the widgets,
because setLoading(false) was called in a callback on loading the
unique store in store ("assignedTagsStore"). Now, the initial selection
is just populated by loading a local store and the selection in
parentWindow is updated by evaluation of the response from
TagZuordnungService, which also saves another server round-trip.