Fix editable state of solldatum* fields in Probe form
Since rev. 2d3ad82f, the readOnly config of all fields
in the form is set, regardless of original setting.
Since for the affected fields, readOnly was intended as
a permanent setting, make them a displayfield, which is
meant for such purpose. Since the fields should stay
selectable and a Date object has to be rendered here,
make SelectableDisplayField a descendant of DisplayField,
leveraging its renderer config.