Move validation to validator functions
This gives the user the chance (at least sometimes) to see why the
save button is disabled. Unfortunately, I could not convince ExtJS
to also validate the coordinate fields after the first combobox
has been edited. This leads to the coordinate fields not being marked
invalid, if e.g. the Netzbetreiber combobox is edited first and
afterwards a Koordinatenart is chosen.