Avoid duplicate config using displayField and valueField
Since valueField defaults to the value of displayField, there's no
need to set both to the same value. In case of the Koordinatenart
widget, this even led to the Ort form being dirty on the first
time such a form was opened after loading the application.
While at it, optimize the choice of settings a bit with regard
to searching comboboxes by typing.