Fix filtering by typing in poi combobox
Fixes a regression of rev. edf904549, which did not take into account
that editable was set to the ExtJS default because the widget set it
to the opposite. Now, the default is left untouched in the widget
definition instead of setting it in all places where the widget is used.
While at it, further simplify widget definition:
_ Fix store ID and simplify initialization
_ Don't unnecessarily call sort() on a store that is configured
  sortOnLoad: true