Generalize record windows setMessages function
The function is now moved to the base class and sets messages for all
child instances of LadaForm and Fieldset. To match validation messages
from the response with the components, the component names are now set
to the expected validation message key, e.g. geolocats for sample
geolocat fieldset messages.