HtmlEncode renderer results.
The documentation of

https://docs.sencha.com/extjs/6.2.0/modern/Ext.grid.Grid.html

says:

» Constructing HTML blocks in code (even in a template) is a common cause
of security problems such as XSS attacks «

which hints to the necessity for renderer functions to ensure saftey.

This commit encodes the result of renderer functions to improve safety.
Excluded where functions where things like dates are processed by other
functions and where safety is ensured otherwise.