Rename refId to JobId.
In order to improve server side OpenAPI documentation the answer of
various endpoints regarding up-/downloading was changed from returning
an object containing a "refId" to a more meaningful member "jobId".
Because this ID references a (up-) download job.
As a consequence there is a change on the client necessary to deal
with this new answer. This is reflected in this commit.