Adjusted client to reflect JobStatus changes on the server.
For clarification reasons we decided on the server side to
serialize JobStatus.status no longer with a separate serializer to lower case entries.
As a result the client has to be changed too to reflect that.

To stay compatible to the print service a simple lowercasing in the
regarding components was chosen.