Set current status of StatusKombi-Button after loading new record.
A bug occured when editing a measurement and saving it:
After the measurement was "reloaded" the formpanel was re-enabled with
`formPanel.setLoading(false);`. This inadvertedly set the state of the
"Status ändern" button to enabled - as a side-effect.

In order to set the state correctly after re-enabling the form we now
set the disabled-state accordingly to `!newRecord.get("statusEdit")`.
Because it is a disable-state `!` is necessary.