Fix display of filtered Tags for query.
Status quo ante was that the code distinguished between ids of assignable
and unassignable Tags - which was wrong in the first place. Then there were
two places where the respective tags should be displayed:

1) Area of unassignable tags
2) Area of assignable tags

The assumption was that every tag which is not assignable should be
unassignable. Since the code is not dealing with the tags itself but has
values which reference a tag there is the possibility that there is an
id referenced which isn't present in the according store.

Say you have a list which contains one element, e.g. [6] and no entry in
the according store references a tag with the id of 6 the display will
be corrupted.

a) it will display the field of unassignable tags, which is obviously
empty

b) it tries to render a tag which is actually not present.

To remedy the situation we distinguish between

- assignable
- unassignable
- unavailable tags

Unassignable tags land as before in the area for unassignable tags.
Assignable tags land as before in the area for assignable tags.

For unavailable tags there are pseudo-tags ("gelöscht") created and
added to the store. Those are displayed alongside the assignable tags.
They indicate that the referring tag is no longer there and allow
the user to deselect the pseudo-tag.