Fix bug where current statusKombi is not shown correctly.
This is a followup commit to 7cf8ab46 where the adjustments alongside
StatusProt/StatusKombi were started.

The situation was that the setValue() function of the StatusKombi-widget
used an id of a StatusProt to retrieve said StatusProt and determine
the kombi to be displayed. Due to changes this id could not be
determined and thus undefined was used.

This commit:

- unifies the determination of statuskombis from a given record by introducing
  a StatusProtMixin which provides a function to do so.
- cleans the rendering of the value up such that only a given value is
  rendered. The retrieval violoates SRP and is removed.
- Removes code duplication in the grid of Messung where the exact same
  function was needed by leveraging the newly introduced mixin.