Reload measm after creation.
After creating a new measm from the values set in the form the object
received from the server contains a correctly formed StatusProt object.
This object is correctly parsed but doesn't get correctly initialized as
an associated model.

Therefore I propose a workaround of reloading the recently created
record and work with the fully initialized measm object from the result
of this request.

This is not optimal - it involves another roundtrip to the server - but does the trick.