Substitute reloading with more graceful parsing.
DetermineKombi tries to parse the StatusKombi either by
retrieving the StatusProt associated. In case that ExtJS messed up the
initialization of a StatusProt it puts the according object as a node on
its measm object where it could be retrieved via get("statusProt").
In either case the statusMpId is necessary to retrieve the Kombi.
So it is either retrieved from the ExtJs object or the JS object.

If statusProt could not be determined in either way, the error is logged
to the console.